fix back bug

+++ wow/src/main/java/com/hqz/wow/service/CouponService.java
@@ -2,10 +2,12 @@ package com.hqz.wow.service;
 
 import com.hqz.wow.entity.CouponEntity;
 
+import java.math.BigInteger;
+
 public interface CouponService {
     CouponEntity getCouponById(int couponId);
 
     boolean checkCouponDate(int couponId);
 
-    boolean checkCouponExists(int couponId);
+    boolean checkCouponExists(BigInteger couponId);
 }
